requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
PyMuPDF>=1.23.7  # fitz
pytesseract>=0.3.10
Pillow>=10.1.0
PyPDF2>=3.0.1
python-docx>=1.0.1
pandas>=2.1.3
PyYAML>=6.0.1
python-magic>=0.4.27
